pyproject.toml
[tool.poetry]
name = "tap-google-drive"
version = "0.1.0"
description = "A Meltano tap for syncing CSV files from Google Drive in magic folder mode"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "tap_google_drive"}]

[tool.poetry.dependencies]
python = "^3.9"
singer-sdk = "^0.34.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.0.0"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pre-commit = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
